Created test_utils module, refactored and moved the HotMusig2Cosigner in it
