+++ app/bitcoin/src/test_utils/mod.rs
@@ -1,0 +1,1 @@
+pub mod musig;
